+++ feature/album/src/main/kotlin/com/igorwojda/showcase/feature/album/presentation/screen/albumlist/AlbumListViewModel.kt
@@ -6,10 +6,10 @@ import androidx.lifecycle.viewModelScope
 import com.igorwojda.showcase.feature.album.domain.model.Album
 import com.igorwojda.showcase.feature.album.domain.usecase.GetAlbumListUseCase
 import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.Action
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.UiState
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.UiState.Content
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.UiState.Error
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.UiState.Loading
+import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.AlbumListUiState
+import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.AlbumListUiState.Content
+import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.AlbumListUiState.Error
+import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.AlbumListUiState.Loading
 import com.igorwojda.showcase.feature.base.domain.result.Result
 import com.igorwojda.showcase.feature.base.presentation.nav.NavManager
 import com.igorwojda.showcase.feature.base.presentation.viewmodel.BaseAction
@@ -22,7 +22,7 @@ internal class AlbumListViewModel(
     private val savedStateHandle: SavedStateHandle,
     private val navManager: NavManager,
     private val getAlbumListUseCase: GetAlbumListUseCase,
-) : BaseViewModel<UiState, Action>(Loading) {
+) : BaseViewModel<AlbumListUiState, Action>(Loading) {
     companion object {
         const val DEFAULT_QUERY_NAME = "Jackson"
         private const val SAVED_QUERY_KEY = "query"
@@ -70,26 +70,26 @@ internal class AlbumListViewModel(
 //        navManager.navigate(navDirections)
     }
 
-    internal sealed interface Action : BaseAction<UiState> {
+    internal sealed interface Action : BaseAction<AlbumListUiState> {
         class AlbumListLoadSuccess(
             private val albums: List<Album>,
         ) : Action {
-            override fun reduce(state: UiState) = Content(albums)
+            override fun reduce(state: AlbumListUiState) = Content(albums)
         }
 
         object AlbumListLoadFailure : Action {
-            override fun reduce(state: UiState) = Error
+            override fun reduce(state: AlbumListUiState) = Error
         }
     }
 
     @Immutable
-    internal sealed interface UiState : BaseState {
+    internal sealed interface AlbumListUiState : BaseState {
         data class Content(
             val albums: List<Album>,
-        ) : UiState
+        ) : AlbumListUiState
 
-        object Loading : UiState
+        object Loading : AlbumListUiState
 
-        object Error : UiState
+        object Error : AlbumListUiState
     }
 }
